+++ app/ide-desktop/lib/content/src/index.ts
@@ -934,6 +934,7 @@ async function runEntryPoint(config: Config) {
     //initCrashHandling()
     style_root()
     printScamWarning()
+    /// Only hide logs in production, but show them when running a development version.
     if (!Versions.isDevVersion()) {
         hideLogs()
     }
